(chore) Reorganize folder structure
