+++ .travis.yml
@@ -1,7 +1,7 @@
 language: node_js
-node_js: 
+node_js:
   - "stable"
 install: npm install
-cache: 
+cache:
   directories:
     - "node_modules"
